Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Running npm install raised two moderate vulnerabilities. Is that a known issue?
> 
> 
> $ npm install
> npm WARN optional SKIPPING OPTIONAL DEPENDENCY: fsevents@1.2.4 (node_modules/fsevents):
> npm WARN notsup SKIPPING OPTIONAL DEPENDENCY: Unsupported platform for fsevents@1.2.4: wanted {"os":"darwin","arch":"any"} (current: {"os":"linux","arch":"x64"})
> 
> audited 6699 packages in 10.384s
> found 2 moderate severity vulnerabilities
>   run `npm audit fix` to fix them, or `npm audit` for details

Yeah, it a vulnerability in one of the development dependencies (rollup-plugin-node-builtins) and it doesn't seem to affect any of us (not will go to production). I need the plugins and I couldn't in a timely manner find a replacement (or fix it myself). Could we merge this and create an issue to solve these vulnerability issues? It can take some time to fix it.

Patch-set: 2
Reviewer: Thiago Cardoso <1013086@d5d70762-12d0-45a1-890d-524b12d3f735>