Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

We should also catch ping timeout errors. When ping timeout happens, a new Error('Ping Timeout') is received in the .on('error') handlers (please check the library to see if it is exactly this.

Unfortunately in this case we should determine the error by comparing with the error message :( (or we can add an error code to the error in the library).

Patch-set: 1
Reviewer: Gerrit User 1013086 <1013086@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1