Fix compatibility with browser in reconnection
This patch removes functions that aren't compatible with browser API.

Use the constructor requested in browser library with 3 parameters:
the URI, the protocol and the websocket options. Instead of using the
node constructor with 2 parameters: URI and websocket option.

And remove the function removeAllListener that is not available in
browser API.

Change-Id: I23142c450cecad4c1b2baa5d7be1833e8eafdaf4
Reviewed-on: https://review.gerrithub.io/c/CESARBR/knot-cloud-websocket/+/451664
Reviewed-by: Paulo Serra Filho <ptvsf@cesar.org.br>
Reviewed-by: Thiago Cardoso <thiago.figueredo@cesar.org.br>
Tested-by: Thiago Cardoso <thiago.figueredo@cesar.org.br>